requirements.txt
moviepy==1.0.3
numpy==2.2.6
openai-whisper==20250625
Pillow==10.2.0
requests==2.32.4
tqdm==4.67.1
torch==2.7.1 